MIPS: fix argv & envp format for MIPS64 builds
The entries in the argv & envp arrays are expected to be 32 bit
addresses. This is the ABI expected by the Linux kernel and it is the
only way to go if a 64 bit U-Boot is to be able to boot a 32 bit Linux
kernel (or vice versa). The U-Boot code generating those arrays was
previously generating an array of native-sized pointers. Generate an
array of 32 bit integers instead.

Signed-off-by: Paul Burton <paul.burton@imgtec.com>