libfdt: allow non-64b aligned memreserve entries
Although the memreserve entries in an FDT are 64 bit aligned relative to
the start of the FDT, we cannot know that the FDT as a whole will be 64
bit aligned in memory. This is most obviously the case when using a FIT
image, where the alignment of the included DTB cannot be specified. On
systems where 64 bit accesses to non-64 bit aligned addresses are not
allowed, this causes problems when accessing the memreserve sections
since this occurs prior to relocating the FDT. Copy the 64 bit values to
suitably aligned on-stack variables to access them in order to avoid
this issue.

Signed-off-by: Paul Burton <paul.burton@imgtec.com>