MIPS: Support for EVA boot
On boot, set the memory map to legacy mode using the SegCtl registers.
This is always done if they are present such that one u-boot image will
boot on either configuration.

Further work would be required if the BEV or physical memory map
prevents the legacy virtual memory map working for a platform.

[paul.burton@imgtec.com:
  - Forward port.
  - Fill delay slots.
  - Conditionalise config1, config2 checks on pre-R6.]

Signed-off-by: Matt Redfearn <matt.redfearn@imgtec.com>