test/py: Encode/decode strings for stdio
When reading stdin python 3.x will give us byte arrays, and when writing
stdout or stderr it will expect byte arrays. In order to insulate the
rest of the code from this difference, call encode or decode at
appropriate points when reading or writing stdio files. This works fine
on python 2.x too.

Signed-off-by: Paul Burton <paul.burton@imgtec.com>
Reviewed-by: Stephen Warren <swarren@nvidia.com>