dtoc: Handle byte-array fdt32_to_cpu case
fdt_to_cpu() can be called with val being a bytes object, not always a
string. This leads to issues when trying to encode the string, for
example when building sandbox_spl_defconfig:

  Traceback (most recent call last):
    File "./tools/dtoc/dtoc", line 380, in <module>
      plat.GenerateTables()
    File "./tools/dtoc/dtoc", line 325, in GenerateTables
      vals.append(self.GetValue(prop.type, val))
    File "./tools/dtoc/dtoc", line 143, in GetValue
      return '%#x' % fdt_util.fdt32_to_cpu(value)
    File "/home/pburton/src/u-boot-up2/tools/dtoc/fdt_util.py", line 27,
in fdt32_to_cpu
      val = val.encode('raw_unicode_escape')
  AttributeError: 'bytes' object has no attribute 'encode'

Fix this by only attempting to encode when actually called with a
string.

Signed-off-by: Paul Burton <paul.burton@imgtec.com>