dm: Support for cache-coherent DMA
Some systems include a subset of devices for which DMA is coherent with
CPU caches. When we have such cache-coherent devices perform DMA we need
to skip cache maintenance operations around the DMA, which may otherwise
be destructive as we inadvertently discard data through cache
invalidations.

This patch adds a flag, DM_FLAG_DMA_CACHE_COHERENT, which can be used to
indicate that a device performs cache-coherent DMA and sets this flag
for devices which have a "dma-coherent" property in device tree. This
property matches similar support in Linux.

New dev_dma_cache_writeback() & dev_dma_cache_invalidate() functions are
introduced which perform the appropriate cache maintenance operations
only when the DM_FLAG_DMA_CACHE_COHERENT flag is not set. The intention
is that drivers begin to migrate to calling these functions instead of
directly calling flush_dcache_range() or invalidate_dcache_range(), and
therefore gain support for systems in which DMA is cache-coherent.

Signed-off-by: Paul Burton <paul.burton@imgtec.com>