boston: Enable display command
Enable the display command, allowing a user to display an arbitrary
message on the LCD display present on the MIPS Boston development board.

We implement the display_set & display_putc functions which back the
display command by buffering the content of the LCD & updating the LCD
using the existing lowlevel_display function.

Signed-off-by: Paul Burton <paul.burton@imgtec.com>