boston: Default to bootdelay=-1
On MIPS Boston boards autoboot can be useful if the board needs to boot
known software every time, for example if it's being used for demo
purposes. In general though the boards are used for development & debug
so it makes sense to default to bootdelay=-1 to disable the auto-boot
feature.

Signed-off-by: Paul Burton <paul.burton@imgtec.com>