net: Exit from eth_rx() if the net_state becomes NETLOOP_FAIL
When net_state is set to NETLOOP_FAIL we are attempting to exit the
network loop due to an error such as a file not found or a permissions
error from a TFTP server. However eth_rx() loops & continues to loop
without regard for net_state, meaning that after net_state becomes
NETLOOP_FAIL we continue to process packets rather than exiting the
network loop & returning to the shell.

This can lead to the rather confusing state in which U-Boot claims that
it's not going to retry an operation but then continues to attempt to
process packets & spams the log with:

  TFTP error: 'File not found' (1)
  Not retrying...

  TFTP error: 'File not found' (1)
  Not retrying...

  TFTP error: 'File not found' (1)
  Not retrying...

Fix this by checking net_state in eth_rx() & breaking out of its loop
when the state becomes NETLOOP_FAIL.

Signed-off-by: Paul Burton <paul.burton@imgtec.com>