Fix u-boot for a big endian CPU with little endian peripherals
Fix u-boot serial, ethernet and pcie drivers to operate in a big
endian CPU with little endian peripherals. If the CPU is in big
endian mode, byte swap the data read from or written to little
endian peripherals where needed.

Signed-off-by: Hassan Naveed <hassan.naveed@imgtec.com>

---
Some subtle merge fixes by ZubairLK
And a few other fixes in the patch.
The DDR3 swap in lowlevelinit.S was of k1. The right reg was t1.
The #def in lowlevelinit.S for little endian needed a nop.