MIPS: Become coherent when L2 isn't bypassed
When the L2 cache is not bypassed during L1 init, we will have already
initialised the L2 first & so skip over its initialisation &
unbypassing. We do still need to become cached & coherent though in
order for the system to perform well & for IOCUs to function. The branch
in the non-bypassed case was skipping a little too much - have it branch
to the code handling the CCA & coherence to fix this.

Signed-off-by: Paul Burton <paul.burton@imgtec.com>