MIPS: sead3: Add device tree file
This device tree file is copied from Linux with the following changes.

- cpu node
- ehci register offset fix
- dm-pre-reloc in uart node.

Diff.

 	cpus {
 		cpu@0 {
-			compatible = "mti,mips14KEc", "mti,mips14Kc";
+			device_type = "cpu";
+			compatible = "cpu-mips";
+			reg = <0>;
+			u-boot,dm-pre-reloc;
 		};
 	};

 	ehci@1b200000 {
 		compatible = "generic-ehci";
-		reg = <0x1b200000 0x1000>;
+		reg = <0x1b200100 0x1000>;
...

uart0: uart@1f000900 {
...
+		u-boot,dm-pre-reloc;
...

Signed-off-by: Zubair Lutfullah Kakakhel <Zubair.Kakakhel@imgtec.com>

---

Not entirely sure if best practise is to remove unused nodes in u-boot
e.g. gic. Or leave them as is. Personally, I'd prefer keeping the dts
files in sync as much as possible.