MIPS: mips-relocs: Ensure we shrink the correct PT_LOAD segment
After emitting relocation information into the .rel section of the
U-Boot ELF, the mips-relocs tool shrinks that section & the PT_LOAD
segment that it is found within. This allows the u-boot.bin binary to be
shrunk to minimal size, rather than leaving space that was reserved at
link time wasted.

Some toolchains produce multiple PT_LOAD segments, and the mips-relocs
tool was not prepared to handle this. It would shrink the first PT_LOAD
segment it found regardless of whether the .rel section is found within
it, effectively discarding the wrong data. Fix this by checking whether
each PT_LOAD segment contains the .rel section, only shrinking the one
that does.

Signed-off-by: Paul Burton <paul.burton@imgtec.com>