dm: syscon: Scan children after binding a syscon
A common pattern in device trees is to have a syscon node whose child
nodes are devices, and which implicitly reference the syscon by way of
the parent-child relationship rather than using a phandle.

In order for this to work we need to scan nodes beneath syscon nodes, so
do this by calling dm_scan_fdt_dev after binding a syscon.

Signed-off-by: Paul Burton <paul.burton@imgtec.com>
Cc: Simon Glass <sjg@chromium.org>